AI-service /requirements.txt
sentence_transformers~=4.1.0
streamlit~=1.33.0
chromadb~=1.0.5
sentence-transformers  # For embedding text into vectors
Flask~=3.0.3
flask-cors~=4.0.0
openai~=1.75.0
reflection~=0.1.1
pandas~=2.2.1
python-dotenv~=1.1.0
requests~=2.31.0
PyMySQL~=1.1.0
pymongo~=4.12.1
numpy~=1.26.2
